+++ back/server/middlewares/req-id-setter.js
@@ -1,0 +1,8 @@
+const { v4: uuid } = require('uuid');
+
+const reqIdSetter = (req, _, next) => {
+  req.id = uuid();
+  next();
+};
+
+module.exports = reqIdSetter;
